fix(country-mode): split client-safe types out of country-rule-alerts (W4.21)
The admin AlertsEditor (a "use client" component) imports types from
@/lib/country-rule-alerts. With the supabase-server fetcher in the same
file, the client bundle pulled in lib/supabase/server.ts → next/headers
→ Turbopack build failure ("only available in Server Components").

Split the module the same way intent-context.ts /
intent-context-server.ts is:

- lib/country-rule-alerts.ts now holds zod schemas + types only —
  client-safe.
- lib/country-rule-alerts-server.ts holds getActiveAlertsForCountry()
  with the supabase import.

Public API route + lib test (server fetcher) re-pointed to the new
server module. Schema/type tests stay on the client-safe module.
